+++ lib/views/dashboard.py
@@ -11,7 +11,7 @@ from app import app, db
 from lib.webmodels import ScoringLog
 from lib.webconfig import FOLDERS, ADMIN_USERS
 from lib.webforms import EditProjectNameForm
-from lib.webmodules import DashboardLog
+from lib.webmodules.webutils import DashboardLog
 
 @app.route('/')
 def index():
